Edit create, delete and edit views to use JWTs to authenticate users. Use localstorage to store userId when a user logs in.

+++ src/routes/deletePost.jsx
@@ -6,11 +6,20 @@ export default function DeletePost() {
 
   const handleDelete = async () => {
     try {
-      const response = await fetch(`http://localhost:3000/posts/${postId}`, { 
+      const token = localStorage.getItem('token'); // Retrieve token from local storage
+      if (!token) {
+        console.error('No token found');
+        localStorage.removeItem('token');
+        navigate('/')
+        return;
+      }
+
+      const response = await fetch(`http://localhost:3000/posts/${postId}`, {
         method: 'DELETE',
         headers: {
           'Content-Type': 'application/json',
-        }
+          'Authorization': `Bearer ${token}`, // Attach token to the request
+        },
       });
 
       const data = await response.json();
